Fix hyphenation of unit and name
Rather than providing explicit hints for hyphenation to work
around the fact that parboxes do not hyphenate the first word,
add a zero-length space in front of the unit and name so that
hyphenation is enabled.

This requires a new \unitname command to allow the zer-width space
to be inserted.